+++ src/reader/core.py
@@ -356,8 +356,11 @@ class Reader:
         _enable_search: bool = False,
         _called_directly: bool = True,
     ):
+        #: The :class:`~reader._storage.Storage` instance used by this reader.
         self._storage = _storage
+        #: The :class:`~reader._search.Search` instance used by this reader.
         self._search = _search
+        #: The :class:`~reader._parser.Parser` instance used by this reader.
         self._parser = _parser
 
         self._reserved_name_scheme = _reserved_name_scheme